aggregation_search/requirements.txt
transformers
torch
torchvision
torchaudio
faiss-cpu
numpy
streamlit
streamlit-folium
ortools
opencv-python-headless
streamlit-image-coordinates
